\name{MCRestimateMerge}
\alias{MCRestimateMerge}
\title{Merging objects of class MCRestimate}
\description{MCRestimateMerge merges objects of S3 class
  \code{MCRestimate} See vignette MCRestimateBlueprint.Rnw for details.}

\usage{
MCRestimateMerge(MCRestimateList)}
\arguments{
  \item{MCRestimateList}{list of objects of S3 class \code{MCRestimate}}}

\value{An objects of S3 class \code{MCRestimate}}

\author{Markus Ruschhaupt \url{mailto:m.ruschhaupt@dkfz.de}}

\seealso{\code{\link{MCRestimate}}}


\examples{## please have a look at the vignette in the scripts folder}

\keyword{}
