\name{additional function for PLR}
\alias{plr.fit}
\alias{bindev} 
\title{internally functions used by PLR.}
\description{These functions are only used internally by the function
  PLR and should not be used by a user.}

\usage{}
\arguments{}



\value{}

\author{Ulrich Mansmann}

\examples{}

\keyword{}
