\name{whatiscorrect}
\alias{whatiscorrect} 
\title{internally used function}
\description{This function is only used internally by the functions
  plot.MCRestimate and MCRestimate and should not be used by a user.}

\usage{}

\arguments{}

\value{}

\author{Markus Ruschhaupt}

\examples{}

\keyword{}
